Refactor  - Add/rename namespaces to align with the endpoint names
