+++ samples/sql-persistence/saga-rename/SqlPersistence_9/EndpointVersion1/MyReplySagaVersion1.cs
@@ -2,52 +2,50 @@
 using System.Threading.Tasks;
 using NServiceBus;
 using NServiceBus.Logging;
+using Shared;
 
 #region replySaga1
+namespace EndpointVersion1;
 
-namespace MyNamespace1
+public class MyReplySagaVersion1 :
+    Saga<MyReplySagaVersion1.SagaData>,
+    IAmStartedByMessages<StartReplySaga>,
+    IHandleMessages<Reply>
 {
-    public class MyReplySagaVersion1 :
-        Saga<MyReplySagaVersion1.SagaData>,
-        IAmStartedByMessages<StartReplySaga>,
-        IHandleMessages<Reply>
+    readonly static ILog log = LogManager.GetLogger<MyReplySagaVersion1>();
+
+    protected override void ConfigureHowToFindSaga(SagaPropertyMapper<SagaData> mapper)
     {
-        readonly static ILog log = LogManager.GetLogger<MyReplySagaVersion1>();
+        mapper.MapSaga(saga => saga.TheId)
+            .ToMessage<StartReplySaga>(msg => msg.TheId)
+            .ToMessage<Reply>(msg => msg.TheId);
+    }
 
-        protected override void ConfigureHowToFindSaga(SagaPropertyMapper<SagaData> mapper)
+    public Task Handle(StartReplySaga message, IMessageHandlerContext context)
+    {
+        var sendOptions = new SendOptions();
+        sendOptions.RouteToThisEndpoint();
+        sendOptions.DelayDeliveryWith(TimeSpan.FromSeconds(10));
+
+        var request = new Request
         {
-            mapper.MapSaga(saga => saga.TheId)
-                .ToMessage<StartReplySaga>(msg => msg.TheId)
-                .ToMessage<Reply>(msg => msg.TheId);
-        }
+            TheId = message.TheId
+        };
 
-        public Task Handle(StartReplySaga message, IMessageHandlerContext context)
-        {
-            var sendOptions = new SendOptions();
-            sendOptions.RouteToThisEndpoint();
-            sendOptions.DelayDeliveryWith(TimeSpan.FromSeconds(10));
+        log.Warn("Saga started. Sending Request");
 
-            var request = new Request
-            {
-                TheId = message.TheId
-            };
+        return context.Send(request, sendOptions);
+    }
 
-            log.Warn("Saga started. Sending Request");
+    public Task Handle(Reply reply, IMessageHandlerContext context)
+    {
+        // throw only for sample purposes
+        throw new Exception("Expected Reply in MyReplySagaVersion2. EndpointVersion1 may have been incorrectly started.");
+    }
 
-            return context.Send(request, sendOptions);
-        }
-
-        public Task Handle(Reply reply, IMessageHandlerContext context)
-        {
-            // throw only for sample purposes
-            throw new Exception("Expected Reply in MyReplySagaVersion2. EndpointVersion1 may have been incorrectly started.");
-        }
-
-        public class SagaData :
-            ContainSagaData
-        {
-            public Guid TheId { get; set; }
-        }
+    public class SagaData : ContainSagaData
+    {
+        public Guid TheId { get; set; }
     }
 }
