+++ samples/sql-persistence/saga-rename/SqlPersistence_9/EndpointVersion2/MyTimeoutSagaVersion2.cs
@@ -2,42 +2,42 @@
 using System.Threading.Tasks;
 using NServiceBus;
 using NServiceBus.Logging;
+using Shared;
 
 #region timeoutSaga2
-namespace MyNamespace2
+namespace EndpointVersion2;
+
+public class MyTimeoutSagaVersion2 :
+    Saga<MyTimeoutSagaVersion2.SagaData>,
+    IAmStartedByMessages<StartTimeoutSaga>,
+    IHandleTimeouts<SagaTimeout>
 {
-    public class MyTimeoutSagaVersion2 :
-        Saga<MyTimeoutSagaVersion2.SagaData>,
-        IAmStartedByMessages<StartTimeoutSaga>,
-        IHandleTimeouts<SagaTimeout>
+    readonly static ILog log = LogManager.GetLogger<MyTimeoutSagaVersion2>();
+
+    protected override void ConfigureHowToFindSaga(SagaPropertyMapper<SagaData> mapper)
     {
-        readonly static ILog log = LogManager.GetLogger<MyTimeoutSagaVersion2>();
+        mapper.MapSaga(saga => saga.TheId)
+            .ToMessage<StartTimeoutSaga>(msg => msg.TheId);
+    }
 
-        protected override void ConfigureHowToFindSaga(SagaPropertyMapper<SagaData> mapper)
-        {
-            mapper.MapSaga(saga => saga.TheId)
-                .ToMessage<StartTimeoutSaga>(msg => msg.TheId);
-        }
+    public Task Handle(StartTimeoutSaga message, IMessageHandlerContext context)
+    {
+        // throw only for sample purposes
+        throw new Exception("Expected StartTimeoutSaga in MyTimeoutSagaVersion1.");
+    }
 
-        public Task Handle(StartTimeoutSaga message, IMessageHandlerContext context)
-        {
-            // throw only for sample purposes
-            throw new Exception("Expected StartTimeoutSaga in MyTimeoutSagaVersion1.");
-        }
+    public Task Timeout(SagaTimeout state, IMessageHandlerContext context)
+    {
+        log.Warn($"Received Timeout from {state.OriginatingSagaType}");
 
-        public Task Timeout(SagaTimeout state, IMessageHandlerContext context)
-        {
-            log.Warn($"Received Timeout from {state.OriginatingSagaType}");
+        MarkAsComplete();
+        return Task.CompletedTask;
+    }
 
-            MarkAsComplete();
-            return Task.CompletedTask;
-        }
-
-        public class SagaData :
-            ContainSagaData
-        {
-            public Guid TheId { get; set; }
-        }
+    public class SagaData :
+        ContainSagaData
+    {
+        public Guid TheId { get; set; }
     }
 }
 #endregion
