+++ samples/sql-persistence/saga-rename/SqlPersistence_9/Shared/Reply.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using NServiceBus;
 
-public class Reply:
-    IMessage
+namespace Shared;
+
+public record Reply : IMessage
 {
     public Guid TheId { get; set; }
     public string OriginatingSagaType { get; set; }
