+++ samples/sql-persistence/saga-rename/SqlPersistence_9/Shared/SagaTimeout.cs
@@ -1,4 +1,6 @@
-public class SagaTimeout
+namespace Shared;
+
+public record SagaTimeout
 {
     public string OriginatingSagaType { get; set; }
 }
