+++ samples/sql-persistence/saga-rename/SqlPersistence_9/Shared/SharedConfiguration.cs
@@ -2,6 +2,8 @@ using System;
 using Microsoft.Data.SqlClient;
 using NServiceBus;
 
+namespace Shared;
+
 public static class SharedConfiguration
 {
     public static void Apply(EndpointConfiguration endpointConfiguration)
@@ -14,8 +16,12 @@ public static class SharedConfiguration
         var subscriptions = persistence.SubscriptionSettings();
         subscriptions.CacheFor(TimeSpan.FromMinutes(1));
 
-        // for SqlExpress use Data Source=.\SqlExpress;Initial Catalog=NsbSamplesSqlPersistenceRenameSaga;Integrated Security=True;Encrypt=false
-        var connectionString = @"Server=localhost,1433;Initial Catalog=NsbSamplesSqlPersistenceRenameSaga;User Id=SA;Password=yourStrong(!)Password;Encrypt=false";
+        // for SqlExpress: 
+        //var connectionString = @"Data Source=.\SqlExpress;Initial Catalog=NsbSamplesSqlPersistenceRenameSaga;Integrated Security=True;Encrypt=false";
+        // for SQL Server:
+        //var connectionString = @"Server=localhost,1433;Initial Catalog=NsbSamplesSqlPersistenceRenameSaga;User Id=SA;Password=yourStrong(!)Password;Encrypt=false";
+        // for LocalDB:
+        var connectionString = @"Data Source=(localdb)\MSSQLLocalDB;Initial Catalog=NsbSamplesSqlPersistenceRenameSaga;Integrated Security=True;Connect Timeout=30;Encrypt=False;Trust Server Certificate=False;Application Intent=ReadWrite;Multi Subnet Failover=False";
 
         persistence.ConnectionBuilder(
             () => new SqlConnection(connectionString));
