+++ samples/sql-persistence/saga-rename/SqlPersistence_9/Shared/StartTimeoutSaga.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using NServiceBus;
 
-public class StartTimeoutSaga:
-    IMessage
+namespace Shared;
+
+public record StartTimeoutSaga : IMessage
 {
     public Guid TheId { get; set; }
 }
